pyproject.toml
[project]
name = "ai-anime-recommendation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.55.0",
]
